[fix] 소셜 로그인했을 때 관리자 권한 설정 줘도 인식못하던거 수정

+++ fitzCode/src/main/java/kr/co/fitzcode/admin/service/SearchLogServiceImpl.java
@@ -13,12 +13,12 @@ public class SearchLogServiceImpl implements SearchLogService {
     private final SearchLogMapper searchLogMapper;
 
     @Override
-    public void saveSearchLog(int userId, String keyword) {
+    public void saveSearchLog(int userId, String keyword) { // 시그니처 맞춤
         log.debug("Attempting to save search log: userId={}, keyword={}", userId, keyword);
         Long count = searchLogMapper.countValidKeyword(keyword);
         log.debug("Keyword validity check: count={}", count);
 
-        if (count != null && count.intValue() > 0) {
+        if (count != null && count > 0) {
             searchLogMapper.insertSearchLog(userId, keyword);
             log.debug("Search log saved: userId={}, keyword={}", userId, keyword);
         } else {
